MNT/ENH: Add FLOAT and FILL options for spreadsheet dataType (#701)
* MNT/ENH: Add FLOAT and FILL options for spreadsheet dataType

This adds two new options to the handled data types, and also raises
an error if we get a datatype that isn't handled. This is important
because otherwise there will be a parameter added without a corresponding
parameterType entry.

* TST: Add tests for Excel to XTCE XML

Add a simple excel file with variable types we want to test and
assert against a known XML file we have produced.