TST: Add new mark to pyproject and re-order fixture dependency
The new mark needs to be registered with pytest to be used, so add
it to the initialization in pyproject.

The fixture using the test_data would try to run before the download
fixture was run. This caused the test data to not be present. To fix
this we can explicitly have the fixture depend on the download data
fixture to make sure it gets run first.