MNT: Factor out the Data_version logic into the common post-processing
The injection of the data version can happen at the CLI call site
rather than having to be passed through to every single instrument.
This means that an instrument doesn't need to specify a version
to produce a product they can just produce a product and the
infrastructure can decide which version that should be.

There is a default version of v001 used when writing a file to cdf
so that tests will produce valid filenames.