MNT/ENH: Decommutate Ultra packets with derived values
The Ultra packet definition has engineering unit conversions, so
we can call the packet_file_to_datasets with `use_derived_value=True`
to get the conversion. These are then l1b products to follow along
with the mission convention. We handle that directly in the l1a
processing by calling the routine again with a different flag. The
metadata for l1a can be re-used, except that FILLVAL, MINVAL, and
MAXVAL are all skipped for l1b because those values try to get
converted to the datatype, but the datatype has changed with enumeration
and conversions from uint8 to str or uint8 to float, etc. For ease
of use, we just skip those metadata fields for now.