MNT: Add CDF attributes for Lo L2 maps
Take advantage of the build_cdf_dataset function to automatically
handle the adding of metadata and variables. However, the Lo code
has already switched to a 2d map dataset, so we don't want to
transform the 1d dataset anymore, but rather allow for our own
processed map dataset to be used as the base.