MNT: Suppress log messages and filter SPP warnings during tests
There are many CDF schema logger.warnings filling the tests that
are not relevant. We can use the command line option to filter
these warnings.

There are also space packet parser warnings that are emitted for
many of our packets that are the wrong size, but that is all we
have so we can ignore them during tests.