FIX: SWE l1b needs to handle all full-cycle case as well
Before we were only setting the local variable if there was a mismatch
in the full cycle length compared to packets. We can ignore this
if-block and just subset the dataset based on full cycle data
regardless and the isel just won't do anything in the case of a
full match.